requirements.txt
Pillow
Flask
tensorflow>=2.0.0
pinecone-client
torchvision
streamlit_lottie
